+++ packages/@modelfusion-cost-calculator/package.json
@@ -47,15 +47,11 @@
     "dev": "tsup --watch src --onSuccess \"pnpm build:dtsMap\"",
     "dist": "pnpm clean:build && pnpm lint && pnpm build"
   },
-  "dependencies": {
-    "@modelfusion/types": "workspace:*"
-  },
   "devDependencies": {
     "@types/node": "18.11.9",
     "@typescript-eslint/eslint-plugin": "^6.1.0",
     "@typescript-eslint/parser": "^6.1.0",
     "eslint": "^8.45.0",
-    "eslint-config-prettier": "9.1.0",
-    "modelfusion": "0.136.0"
+    "eslint-config-prettier": "9.1.0"
   }
 }
