+++ packages/@modelfusion-cost-calculator/src/openai/calculators/completions.ts
@@ -1,8 +1,21 @@
-import {
-  OpenAICompletionModelType,
-  OpenAICompletionResponse,
-  getOpenAICompletionModelInformation,
-} from "@modelfusion/types";
+/**
+ * @see https://openai.com/pricing
+ */
+export const OPENAI_TEXT_GENERATION_MODELS = {
+  "gpt-3.5-turbo-instruct": {
+    promptTokenCostInMillicents: 0.15,
+    completionTokenCostInMillicents: 0.2,
+  },
+};
+
+type OpenAICompletionModelType = keyof typeof OPENAI_TEXT_GENERATION_MODELS;
+
+export type OpenAICompletionResponse = {
+  usage: {
+    prompt_tokens: number;
+    completion_tokens: number;
+  };
+};
 
 export const calculateOpenAICompletionCostInMillicents = ({
   model,
@@ -11,7 +24,7 @@ export const calculateOpenAICompletionCostInMillicents = ({
   model: OpenAICompletionModelType;
   response: OpenAICompletionResponse;
 }) => {
-  const modelInformation = getOpenAICompletionModelInformation(model);
+  const modelInformation = OPENAI_TEXT_GENERATION_MODELS[model];
 
   return (
     response.usage.prompt_tokens *
@@ -20,3 +33,7 @@ export const calculateOpenAICompletionCostInMillicents = ({
       modelInformation.completionTokenCostInMillicents
   );
 };
+
+export const isOpenAICompletionModel = (
+  model: string
+): model is OpenAICompletionModelType => model in OPENAI_TEXT_GENERATION_MODELS;
