+++ packages/@modelfusion-cost-calculator/src/openai/calculators/speech.ts
@@ -1,20 +1,21 @@
-import {
-  OpenAISpeechModelType,
-  OPENAI_SPEECH_MODELS,
-} from "@modelfusion/types";
+// cost per character in millicents
+const SpeechModelCosts = {
+  "tts-1": 1.5, // = 1500 / 1000,
+  "tts-1-hd": 3, // = 3000 / 1000
+};
+
+export type OpenAISpeechModelType = keyof typeof SpeechModelCosts;
 
 export const calculateOpenAISpeechCostInMillicents = ({
   model,
   input,
 }: {
-  model: OpenAISpeechModelType;
+  model: keyof typeof SpeechModelCosts;
   input: string;
 }): number | null => {
-  if (!OPENAI_SPEECH_MODELS[model]) {
+  if (!SpeechModelCosts[model]) {
     return null;
   }
 
-  return (
-    input.length * OPENAI_SPEECH_MODELS[model].costInMillicentsPerCharacter
-  );
+  return input.length * SpeechModelCosts[model];
 };
