+++ packages/@modelfusion-cost-calculator/src/openai/calculators/transcription.ts
@@ -1,15 +1,23 @@
-import {
-  OpenAITranscriptionModelType,
-  OpenAITranscriptionVerboseJsonResponse,
-  OPENAI_TRANSCRIPTION_MODELS,
-} from "@modelfusion/types";
+/**
+ * @see https://openai.com/pricing
+ */
+export const TRANSCRIPTION_MODEL_COSTS = {
+  "whisper-1": 10, // = 600 / 60,
+};
+
+export type OpenAITranscriptionModelType =
+  keyof typeof TRANSCRIPTION_MODEL_COSTS;
+
+export type TranscriptionResponse = {
+  duration: number;
+};
 
 export const calculateOpenAITranscriptionCostInMillicents = ({
   model,
   response,
 }: {
   model: OpenAITranscriptionModelType;
-  response: OpenAITranscriptionVerboseJsonResponse;
+  response: TranscriptionResponse;
 }): number | null => {
   if (model !== "whisper-1") {
     return null;
@@ -17,8 +25,5 @@ export const calculateOpenAITranscriptionCostInMillicents = ({
 
   const durationInSeconds = response.duration;
 
-  return (
-    Math.ceil(durationInSeconds) *
-    OPENAI_TRANSCRIPTION_MODELS[model].costInMillicentsPerSecond
-  );
+  return Math.ceil(durationInSeconds) * TRANSCRIPTION_MODEL_COSTS[model];
 };
