--- packages/@modelfusion-types/README.md
@@ -1,5 +1,0 @@
-# @modelfusion/cost-calculator
-
-## Usage
-
-TBD...
