--- packages/@modelfusion-types/package.json
@@ -1,57 +1,0 @@
-{
-  "name": "@modelfusion/types",
-  "author": "Lars Grammel",
-  "license": "MIT",
-  "keywords": [
-    "openai",
-    "modelfusion",
-    "typescript"
-  ],
-  "homepage": "https://modelfusion.dev/",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/lgrammel/modelfusion"
-  },
-  "bugs": {
-    "url": "https://github.com/lgrammel/modelfusion/issues"
-  },
-  "type": "module",
-  "sideEffects": false,
-  "private": false,
-  "engines": {
-    "node": ">=18"
-  },
-  "publishConfig": {
-    "directory": "dist",
-    "linkDirectory": true
-  },
-  "files": [
-    "./*"
-  ],
-  "exports": {
-    ".": {
-      "types": "./index.d.ts",
-      "import": "./index.js",
-      "require": "./index.cjs"
-    }
-  },
-  "scripts": {
-    "lint": "eslint --ext .ts src",
-    "clean": "rimraf build dist .turbo node_modules",
-    "clean:build": "rimraf build dist",
-    "build": "tsup --dts && pnpm build:copy-files",
-    "build:copy-files": "copyfiles --flat package.json README.md ../../LICENSE dist",
-    "build:dtsMap": "tsc --emitDeclarationOnly --declaration --declarationMap",
-    "dev": "tsup --watch src --onSuccess \"pnpm build:dtsMap\"",
-    "dist": "pnpm clean:build && pnpm lint && pnpm build"
-  },
-  "dependencies": {},
-  "devDependencies": {
-    "@types/node": "18.11.9",
-    "@typescript-eslint/eslint-plugin": "^6.1.0",
-    "@typescript-eslint/parser": "^6.1.0",
-    "eslint": "^8.45.0",
-    "eslint-config-prettier": "9.1.0",
-    "zod": "3.22.4"
-  }
-}
