--- packages/@modelfusion-types/tsconfig.cjs.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "declaration": false,
-    "allowSyntheticDefaultImports": true
-  },
-  "exclude": ["node_modules"]
-}
