--- packages/@modelfusion-types/tsup.config.ts
@@ -1,9 +1,0 @@
-import { defineConfig } from "tsup";
-
-export default defineConfig([
-  {
-    entry: ["src/index.ts"],
-    format: ["cjs", "esm"],
-    sourcemap: true,
-  },
-]);
