+++ packages/modelfusion/src/model-provider/openai/AbstractOpenAICompletionModel.ts
@@ -1,10 +1,4 @@
-import {
-  OpenAICompletionResponseSchema,
-  OpenAICompletionResponse,
-  OpenAICompletionStreamChunk,
-  openaiCompletionStreamChunkSchema,
-} from "@modelfusion/types";
-
+import z from "zod";
 import { FunctionCallOptions } from "../../core/FunctionOptions";
 import { ApiConfiguration } from "../../core/api/ApiConfiguration";
 import { callWithRetryAndThrottle } from "../../core/api/callWithRetryAndThrottle";
@@ -41,6 +35,56 @@ export interface AbstractOpenAICompletionModelSettings
 
   isUserIdForwardingEnabled?: boolean;
 }
+
+export const OpenAICompletionResponseSchema = z.object({
+  id: z.string(),
+  choices: z.array(
+    z.object({
+      finish_reason: z
+        .enum(["stop", "length", "content_filter"])
+        .optional()
+        .nullable(),
+      index: z.number(),
+      logprobs: z.nullable(z.any()),
+      text: z.string(),
+    })
+  ),
+  created: z.number(),
+  model: z.string(),
+  system_fingerprint: z.string().optional(),
+  object: z.literal("text_completion"),
+  usage: z.object({
+    prompt_tokens: z.number(),
+    completion_tokens: z.number(),
+    total_tokens: z.number(),
+  }),
+});
+
+export type OpenAICompletionResponse = z.infer<
+  typeof OpenAICompletionResponseSchema
+>;
+
+export const openaiCompletionStreamChunkSchema = z.object({
+  choices: z.array(
+    z.object({
+      text: z.string(),
+      finish_reason: z
+        .enum(["stop", "length", "content_filter"])
+        .optional()
+        .nullable(),
+      index: z.number(),
+    })
+  ),
+  created: z.number(),
+  id: z.string(),
+  model: z.string(),
+  system_fingerprint: z.string().optional(),
+  object: z.literal("text_completion"),
+});
+
+export type OpenAICompletionStreamChunk = z.infer<
+  typeof openaiCompletionStreamChunkSchema
+>;
 
 /**
  * Abstract completion model that calls an API that is compatible with the OpenAI completions API.
